[tools] fix build error for alinkapp



See merge request !495

+++ tools/cli/dumpsys.c
@@ -123,7 +123,7 @@ uint32_t dumpsys_task_func(char *buf, uint32_t len, int detail)
 
         snprintf(tmpbuf,255,"%-19s%-9s%-5d%-10d%-9u%-9u%-11c\r\n",
                    task_name, cpu_stat[task->task_state - K_RDY], task->prio,
-                   task->stack_size, (unsigned long)free_size, (unsigned long)time_total, yes);
+                   task->stack_size, free_size, (unsigned int)time_total, yes);
 #endif
         safesprintf(printbuf, totallen, offset,tmpbuf);
